remove no-sharding configuration
