+++ nsdb-cluster/src/main/resources/cluster.conf
@@ -126,7 +126,6 @@ nsdb {
   }
 
   sharding {
-    enabled = false
     interval = 1d
   }
