+++ nsdb-cluster/src/multi-jvm/scala/io/radicalbit/nsdb/cluster/actor/MetadataTest.scala
@@ -38,7 +38,6 @@ object MetadataTest extends MultiNodeConfig {
     |  write.scheduler.interval = 15 seconds
     |
     |  sharding {
-    |    enabled = true
     |    interval = 1d
     |  }
     |
