+++ nsdb-cluster/src/test/resources/application.conf
@@ -32,7 +32,6 @@ nsdb{
   }
 
   sharding {
-    enabled = false
     interval = 1d
   }
