+++ nsdb-cluster/src/test/scala/io/radicalbit/nsdb/cluster/coordinator/ReadCoordinatorBehaviour.scala
@@ -969,5 +969,4 @@ trait ReadCoordinatorBehaviour { this: TestKit with WordSpecLike with Matchers =
       }
     }
   }
-
 }
